requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
Deprecated==1.2.14
distlib==0.3.9
Django==5.1.2
django-browser-reload==1.16.0
django-cors-headers==4.5.0
django-extensions==3.2.3
django-filter==24.3
django-google-cloud-tasks==2.17.0
django-ordered-model==3.7.1
django-template-partials==24.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-extensions==0.7.1
drf-kit==1.44.0
factory_boy==3.3.1
Faker==30.6.0
filelock==3.16.1
gcp-pilot==1.29.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-pubsub==2.26.1
google-cloud-scheduler==2.13.5
google-cloud-tasks==2.16.5
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==23.0.0
httplib2==0.22.0
identify==2.6.1
idna==3.10
importlib_metadata==8.4.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
platformdirs==4.3.6
pre_commit==4.0.1
proto-plus==1.24.0
protobuf==5.28.2
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
rsa==4.9
ruff==0.7.1
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.27.0
wrapt==1.16.0
zipp==3.20.2
